requirements.txt
langchain==0.3.23
langchain_community==0.3.21
langchain_core==0.3.51
langchain_huggingface==0.1.2
langchain_writer==0.1.2
python-dotenv==1.1.0
streamlit==1.44.1
streamlit_chat==0.1.1
